+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+**???***
+* Added class to allow the writing of OLE files, which allows for embedded MSG files to be extracted.
+* Fixed minor issues with `utils.inputToMsgPath` (named from `utils.inputToMsgpath`).
+* Renamed `utils.msgpathToString` to `utils.msgPathToString`.
+
 **v0.37.0**
 * [[TeamMsgExtractor #303](https://github.com/TeamMsgExtractor/msg-extractor/issues/303)] Renamed internal variable that wasn't changed when the other instances or it were renamed.
 * [[TeamMsgExtractor #302](https://github.com/TeamMsgExtractor/msg-extractor/issues/302)] Fixed properties missing (a fatal MSG error) raising an unclear exception. It now uses `InvalidFileFormatError`.